+++ CHANGELOG.md
@@ -33,6 +33,7 @@ TIME_CHOICES = (
 
 ###### TEMPLATES
 - people job_function (which it was used in teojakob for example) was substituted with property 'units'. You now have to add categories (with identifier='units') and tag th person with it. this allowes us to categories people without having to maintain both fields 'unit' and categories
+- locations app has now also `grid_static` as a default template option
 
 ###### URLS